refactor(server/events): extract makeCtrlMetaObserver to observers/ctrl-meta.ts
Move document-meta (open/close/switch) CTRL_ROOM observer into a standalone
factory. Local lastActiveDocId / lastOpenDocIds become closure state per call.
attachCtrlObservers pushes the returned cleanup. Remove now-unused
Y_MAP_DOCUMENT_META and getOpenDocs imports from queue.ts.
